requirements.txt
praw==7.7.1
python-dotenv==1.0.0
pandas==2.1.1
streamlit==1.29.0
google-generativeai==0.3.2

